# Fix use of thread-specific distributions in random parameters