--- a/extras/ci_build.sh
+++ b/extras/ci_build.sh
@@ -84,6 +84,7 @@ if [ "$xNEST_BUILD_TYPE" = "STATIC_CODE_ANALYSIS" ]; then
 
     echo "MSGBLD0070: Retrieving changed files."
     file_names=$CHANGED_FILES
+    echo "MSGBLD0071: $file_names"
 
     # Note: uncomment the following line to static check *all* files, not just those that have changed.
     # Warning: will run for a very long time
@@ -104,6 +105,7 @@ if [ "$xNEST_BUILD_TYPE" = "STATIC_CODE_ANALYSIS" ]; then
     CPPCHECK=cppcheck
     CLANG_FORMAT=clang-format
     PEP8=pycodestyle
+    PYCODESTYLE_IGNORES="E121,E123,E126,E226,E24,E704,W503,W504"
 
     # Perform or skip a certain analysis.
     PERFORM_VERA=true
@@ -126,7 +128,8 @@ if [ "$xNEST_BUILD_TYPE" = "STATIC_CODE_ANALYSIS" ]; then
     ./extras/static_code_analysis.sh "$RUNS_ON_CI" "$INCREMENTAL" "$file_names" "$NEST_VPATH" \
     "$VERA" "$CPPCHECK" "$CLANG_FORMAT" "$PEP8" \
     "$PERFORM_VERA" "$PERFORM_CPPCHECK" "$PERFORM_CLANG_FORMAT" "$PERFORM_PEP8" \
-    "$IGNORE_MSG_VERA" "$IGNORE_MSG_CPPCHECK" "$IGNORE_MSG_CLANG_FORMAT" "$IGNORE_MSG_PYCODESTYLE"
+    "$IGNORE_MSG_VERA" "$IGNORE_MSG_CPPCHECK" "$IGNORE_MSG_CLANG_FORMAT" "$IGNORE_MSG_PYCODESTYLE" \
+    "$PYCODESTYLE_IGNORES"
 
     exit $?
 fi