--- a/nestkernel/connection_creator_impl.h
+++ b/nestkernel/connection_creator_impl.h
@@ -94,7 +94,7 @@ ConnectionCreator::connect_to_target_( Iterator from,
     }
     iter->first.get_vector( source_pos );
 
-    if ( without_kernel or rng->drand() < kernel_->value( rng, source_pos, target_pos, source ) )
+    if ( without_kernel or rng->drand() < kernel_->value( rng, source_pos, target_pos, source, tgt_ptr ) )
     {
       for ( size_t indx = 0; indx < synapse_model_.size(); ++indx )
       {
@@ -103,8 +103,8 @@ ConnectionCreator::connect_to_target_( Iterator from,
           tgt_thread,
           synapse_model_[ indx ],
           param_dicts_[ indx ][ tgt_thread ],
-          delay_[ indx ]->value( rng, source_pos, target_pos, source ),
-          weight_[ indx ]->value( rng, source_pos, target_pos, source ) );
+          delay_[ indx ]->value( rng, source_pos, target_pos, source, tgt_ptr ),
+          weight_[ indx ]->value( rng, source_pos, target_pos, source, tgt_ptr ) );
       }
     }
   }
@@ -418,7 +418,7 @@ ConnectionCreator::fixed_indegree_( Layer< D >& source,
               ++iter )
         {
           iter->first.get_vector( source_pos_vector );
-          probabilities.push_back( kernel_->value( rng, source_pos_vector, target_pos_vector, source ) );
+          probabilities.push_back( kernel_->value( rng, source_pos_vector, target_pos_vector, source, tgt ) );
         }
 
         if ( positions.empty()
@@ -456,8 +456,8 @@ ConnectionCreator::fixed_indegree_( Layer< D >& source,
           positions[ random_id ].first.get_vector( source_pos_vector );
           for ( size_t indx = 0; indx < synapse_model_.size(); ++indx )
           {
-            const double w = weight_[ indx ]->value( rng, source_pos_vector, target_pos_vector, source );
-            const double d = delay_[ indx ]->value( rng, source_pos_vector, target_pos_vector, source );
+            const double w = weight_[ indx ]->value( rng, source_pos_vector, target_pos_vector, source, tgt );
+            const double d = delay_[ indx ]->value( rng, source_pos_vector, target_pos_vector, source, tgt );
             kernel().connection_manager.connect(
               source_id, tgt, target_thread, synapse_model_[ indx ], param_dicts_[ indx ][ target_thread ], d, w );
           }
@@ -495,8 +495,8 @@ ConnectionCreator::fixed_indegree_( Layer< D >& source,
           index source_id = positions[ random_id ].second;
           for ( size_t indx = 0; indx < synapse_model_.size(); ++indx )
           {
-            const double w = weight_[ indx ]->value( rng, source_pos_vector, target_pos_vector, source );
-            const double d = delay_[ indx ]->value( rng, source_pos_vector, target_pos_vector, source );
+            const double w = weight_[ indx ]->value( rng, source_pos_vector, target_pos_vector, source, tgt );
+            const double d = delay_[ indx ]->value( rng, source_pos_vector, target_pos_vector, source, tgt );
             kernel().connection_manager.connect(
               source_id, tgt, target_thread, synapse_model_[ indx ], param_dicts_[ indx ][ target_thread ], d, w );
           }
@@ -548,7 +548,7 @@ ConnectionCreator::fixed_indegree_( Layer< D >& source,
               ++iter )
         {
           iter->first.get_vector( source_pos_vector );
-          probabilities.push_back( kernel_->value( rng, source_pos_vector, target_pos_vector, source ) );
+          probabilities.push_back( kernel_->value( rng, source_pos_vector, target_pos_vector, source, tgt ) );
         }
 
         // A Vose object draws random integers with a non-uniform
@@ -579,8 +579,8 @@ ConnectionCreator::fixed_indegree_( Layer< D >& source,
           ( *positions )[ random_id ].first.get_vector( source_pos_vector );
           for ( size_t indx = 0; indx < synapse_model_.size(); ++indx )
           {
-            const double w = weight_[ indx ]->value( rng, source_pos_vector, target_pos_vector, source );
-            const double d = delay_[ indx ]->value( rng, source_pos_vector, target_pos_vector, source );
+            const double w = weight_[ indx ]->value( rng, source_pos_vector, target_pos_vector, source, tgt );
+            const double d = delay_[ indx ]->value( rng, source_pos_vector, target_pos_vector, source, tgt );
             kernel().connection_manager.connect(
               source_id, tgt, target_thread, synapse_model_[ indx ], param_dicts_[ indx ][ target_thread ], d, w );
           }
@@ -617,8 +617,8 @@ ConnectionCreator::fixed_indegree_( Layer< D >& source,
           ( *positions )[ random_id ].first.get_vector( source_pos_vector );
           for ( size_t indx = 0; indx < synapse_model_.size(); ++indx )
           {
-            const double w = weight_[ indx ]->value( rng, source_pos_vector, target_pos_vector, source );
-            const double d = delay_[ indx ]->value( rng, source_pos_vector, target_pos_vector, source );
+            const double w = weight_[ indx ]->value( rng, source_pos_vector, target_pos_vector, source, tgt );
+            const double d = delay_[ indx ]->value( rng, source_pos_vector, target_pos_vector, source, tgt );
             kernel().connection_manager.connect(
               source_id, tgt, target_thread, synapse_model_[ indx ], param_dicts_[ indx ][ target_thread ], d, w );
           }
@@ -705,7 +705,8 @@ ConnectionCreator::fixed_outdegree_( Layer< D >& source,
       {
         // TODO: Why is probability calculated in source layer, but weight and delay in target layer?
         target_pos_node_id_pair.first.get_vector( target_pos_vector );
-        probabilities.push_back( kernel_->value( grng, source_pos_vector, target_pos_vector, source ) );
+        const auto tgt = kernel().node_manager.get_node_or_proxy( target_pos_node_id_pair.second );
+        probabilities.push_back( kernel_->value( grng, source_pos_vector, target_pos_vector, source, tgt ) );
       }
     }
     else
@@ -752,8 +753,9 @@ ConnectionCreator::fixed_outdegree_( Layer< D >& source,
       std::vector< double > rng_delay_vec;
       for ( size_t indx = 0; indx < weight_.size(); ++indx )
       {
-        rng_weight_vec.push_back( weight_[ indx ]->value( grng, source_pos_vector, target_pos_vector, target ) );
-        rng_delay_vec.push_back( delay_[ indx ]->value( grng, source_pos_vector, target_pos_vector, target ) );
+        const auto tgt = kernel().node_manager.get_node_or_proxy( target_pos_node_id_pairs[ indx ].second );
+        rng_weight_vec.push_back( weight_[ indx ]->value( grng, source_pos_vector, target_pos_vector, target, tgt ) );
+        rng_delay_vec.push_back( delay_[ indx ]->value( grng, source_pos_vector, target_pos_vector, target, tgt ) );
       }
 
       // We bail out for non-local neurons only now after all possible