--- a/nestkernel/parameter.cpp
+++ b/nestkernel/parameter.cpp
@@ -76,7 +76,7 @@ Parameter::apply( const NodeCollectionPTR& nc, const TokenArray& token_array )
           target_pos.size(),
           source_pos.size() ) );
     }
-    auto value = this->value( rng, source_pos, target_pos, *source_layer.get() );
+    auto value = this->value( rng, source_pos, target_pos, *source_layer.get(), nullptr );
     result.push_back( value );
   }
   return result;
@@ -176,7 +176,8 @@ double
 SpatialDistanceParameter::value( RngPtr,
   const std::vector< double >& source_pos,
   const std::vector< double >& target_pos,
-  const AbstractLayer& layer )
+  const AbstractLayer& layer,
+  Node* )
 {
   switch ( dimension_ )
   {
@@ -240,7 +241,8 @@ double
 RedrawParameter::value( RngPtr rng,
   const std::vector< double >& source_pos,
   const std::vector< double >& target_pos,
-  const AbstractLayer& layer )
+  const AbstractLayer& layer,
+  Node* node )
 {
   double value;
   size_t num_redraws = 0;
@@ -250,7 +252,7 @@ RedrawParameter::value( RngPtr rng,
     {
       throw KernelException( String::compose( "Number of redraws exceeded limit of %1", max_redraws_ ) );
     }
-    value = p_->value( rng, source_pos, target_pos, layer );
+    value = p_->value( rng, source_pos, target_pos, layer, node );
   } while ( value < min_ or value > max_ );
 
   return value;
@@ -273,9 +275,10 @@ double
 ExpDistParameter::value( RngPtr rng,
   const std::vector< double >& source_pos,
   const std::vector< double >& target_pos,
-  const AbstractLayer& layer )
+  const AbstractLayer& layer,
+  Node* node )
 {
-  return std::exp( -p_->value( rng, source_pos, target_pos, layer ) * inv_beta_ );
+  return std::exp( -p_->value( rng, source_pos, target_pos, layer, node ) * inv_beta_ );
 }
 
 GaussianParameter::GaussianParameter( const DictionaryDatum& d )
@@ -295,9 +298,10 @@ double
 GaussianParameter::value( RngPtr rng,
   const std::vector< double >& source_pos,
   const std::vector< double >& target_pos,
-  const AbstractLayer& layer )
+  const AbstractLayer& layer,
+  Node* node )
 {
-  const auto dx = p_->value( rng, source_pos, target_pos, layer ) - mean_;
+  const auto dx = p_->value( rng, source_pos, target_pos, layer, node ) - mean_;
   return std::exp( -dx * dx * inv_two_std2_ );
 }
 
@@ -340,10 +344,11 @@ double
 Gaussian2DParameter::value( RngPtr rng,
   const std::vector< double >& source_pos,
   const std::vector< double >& target_pos,
-  const AbstractLayer& layer )
+  const AbstractLayer& layer,
+  Node* node )
 {
-  const auto dx = px_->value( rng, source_pos, target_pos, layer ) - mean_x_;
-  const auto dy = py_->value( rng, source_pos, target_pos, layer ) - mean_y_;
+  const auto dx = px_->value( rng, source_pos, target_pos, layer, node ) - mean_x_;
+  const auto dy = py_->value( rng, source_pos, target_pos, layer, node ) - mean_y_;
   return std::exp( -dx * dx * x_term_const_ - dy * dy * y_term_const_ + dx * dy * xy_term_const_ );
 }
 
@@ -370,9 +375,10 @@ double
 GammaParameter::value( RngPtr rng,
   const std::vector< double >& source_pos,
   const std::vector< double >& target_pos,
-  const AbstractLayer& layer )
+  const AbstractLayer& layer,
+  Node* node )
 {
-  const auto x = p_->value( rng, source_pos, target_pos, layer );
+  const auto x = p_->value( rng, source_pos, target_pos, layer, node );
   return std::pow( x, kappa_ - 1. ) * std::exp( -1. * inv_theta_ * x ) * delta_;
 }