Fixes #2119 by correcting the use of thread-specific distributions in the normal and log-normal random parameters. 

Also adds a test that checks that values from random distribution parameters are invariant with different number of MPI processes and threads, and fixed number of virtual processes.