@hakonsbm 2021-08-20
@heplesser I've updated the implementation, so the SLI test is no longer failing. 

I've also done some simple benchmarks with `timeit`:
```python
n = nest.Create('iaf_psc_alpha', 100000)
p = nest.random.normal()
%timeit n.V_m = p
```
Which results in
| master | this PR |
|-----------|----------|
| 759 ms ± 28.5 ms | 757 ms ± 26.1 ms |

So it seems like the compiler does some optimisations on the division, and performance isn't an issue here.

Although it seems like all tests are passing now, this is not the case. One PyNEST test fails with a segfault (hinted at in [this line](https://github.com/nest/nest-simulator/runs/3379615413#step:16:33746)), but is not reported in the installcheck results. In fact, no PyNEST tests, except for the new MPI ones, are reported there, so it may be that failing PyNEST tests are going unnoticed. I can't see the same problem in other recent PRs, but I don't think I've touched any files that should change this behaviour. @jougs Since you refactored the testsuite recently, could you take a look?