@jougs 2021-08-20
Even though I don't think that this is related to my recent refactoring of the PyNEST testsuite, I can indeed reproduce the segfault locally by running `python install/share/nest/testsuite/pytests/test_nodeParametrization.py`. Here's a backtrace obtained from `gdb`:

```gdb
Test positions specified by parameter ... 
Thread 1 "python" received signal SIGSEGV, Segmentation fault.
nest::VPManager::node_id_to_vp (node_id=<optimized out>, this=<optimized out>)
    at /home/jochen/tmp/nest-simulator/nestkernel/random_generators.h:296
296	  inline result_type operator()( RngPtr g )
(gdb) bt
#0  nest::VPManager::node_id_to_vp (node_id=<optimized out>, this=<optimized out>)
    at /home/jochen/tmp/nest-simulator/nestkernel/random_generators.h:296
#1  nest::NormalParameter::value (this=0x1e363c0, rng=0x1f6fef0, node=0x0)
    at /home/jochen/tmp/nest-simulator/nestkernel/parameter.cpp:105
#2  0x00007ffff4e425c0 in nest::DimensionParameter::get_values (this=this@entry=0x1c16fc0, 
    rng=rng@entry=0x1f6fef0) at /home/jochen/tmp/nest-simulator/nestkernel/parameter.h:1234
#3  0x00007ffff4e4942b in nest::FreeLayer<2>::set_status (this=0x1b03350, d=...)
    at /home/jochen/tmp/nest-simulator/nestkernel/free_layer.h:178
#4  0x00007ffff4e4076e in nest::AbstractLayer::create_layer (layer_dict=...)
    at /home/jochen/tmp/nest-simulator/nestkernel/layer.cpp:168
#5  0x00007ffff4e6476c in nest::create_layer (layer_dict=...)
    at /home/jochen/tmp/nest-simulator/nestkernel/spatial.cpp:118
#6  0x00007ffff4d93fd7 in nest::NestModule::CreateLayer_D_DFunction::execute (this=<optimized out>, 
    i=0xe363e0) at /home/jochen/tmp/nest-simulator/nestkernel/nestmodule.cpp:2145
#7  0x00007ffff4bbbe02 in SLIInterpreter::execute_ (this=0xe363e0, exitlevel=0)
    at /home/jochen/tmp/nest-simulator/sli/interpret.cc:1339
#8  0x00007ffff54e0f7d in __pyx_pf_12pynestkernel_10NESTEngine_8run (__pyx_v_cmd=<optimized out>, 
    __pyx_v_self=0x7ffff7370f70) at /home/jochen/tmp/nest-simulator.build/pynest/pynestkernel.cxx:5689
#9  __pyx_pw_12pynestkernel_10NESTEngine_9run (__pyx_v_self=0x7ffff7370f70, 
    __pyx_v_cmd=<optimized out>) at /home/jochen/tmp/nest-simulator.build/pynest/pynestkernel.cxx:5582
```

Could the `node=0x0` in frame #⁠1 be the culprit here?