@hakonsbm 2021-08-20
@jougs Sorry, I meant the problem with PyNEST tests not showing up in the testsuite results and failing tests not causing the testsuite to fail. The segfault is because of this PR and I'm already working on fixing it.