@niltonlk 2021-08-20
I confirm that @hakonsbm branch solves #2119 

1. fix_normal_random branch 4 MPI w/ 1 OpenMP:
```
$ mpirun -np 4 python3 ./run_microcircuit.py

              -- N E S T --
  Copyright (C) 2004 The NEST Initiative

 Version: fix_normal_random@3dc3d38bb

Interval to compute firing rates: [ 500. 1500.] ms
Mean rates: [0.419 1.997 3.979 4.909 6.449 7.66  0.775 6.953] spikes/s
Standard deviation of rates: [0.733 1.688 2.75  3.306 4.056 4.412 1.082 3.679] spikes/s
```

2. fix_normal_random branch 1 MPI w/ 4 OpenMP:
```
$ mpirun -np 1 python3 ./run_microcircuit.py 

              -- N E S T --
  Copyright (C) 2004 The NEST Initiative

 Version: fix_normal_random@3dc3d38bb

Interval to compute firing rates: [ 500. 1500.] ms
Mean rates: [0.419 1.997 3.979 4.909 6.449 7.66  0.775 6.953] spikes/s
Standard deviation of rates: [0.733 1.688 2.75  3.306 4.056 4.412 1.082 3.679] spikes/s
```

Performance with the 0.1 scaled network seems to be small:

1. 4 MPI w/ 1 OpenMP:

|                                     | NEST 3.0 | fix_normal_random |
|------------------------------|---------------|----------------------------|
|  Total time:                  | 6.579 s     | 6.950 s                    | 
|  Time to initialize:       | 0.064 s     | 0.036 s                     |
|  Time to create:          | 0.074 s     | 0.071 s                     |
|  Time to connect:       | 0.484 s     | 0.498 s                     |
|  Time to presimulate: | 1.559 s     | 1.805 s                     |
|  Time to simulate:      | 3.344 s     | 3.482 s                     |
|  Time to evaluate:      | 1.053 s     | 1.058 s                     |

2. 1 MPI w/ 4 OpenMP:

|                                     | NEST 3.0 | fix_normal_random |
|------------------------------|---------------|----------------------------|
|  Total time:                  | 19.195 s   | 19.542 s                   | 
|  Time to initialize:       | 0.004 s     | 0.004 s                     |
|  Time to create:          | 0.138 s     | 0.138 s                     |
|  Time to connect:       | 1.526 s     | 1.543 s                     |
|  Time to presimulate: | 5.615 s     | 5.683 s                     |
|  Time to simulate:      | 10.965 s   | 11.230 s                    |
|  Time to evaluate:      | 0.946 s     | 0.945 s                     |

Now, why is the OpenMP parallelization taking about 3 times longer than the MPI parallelization? More surprisingly, OpenMP parallelization was slower than no parallelization at all?:

3. 1 MPI w/ 1 OpenMP:

|                                     | NEST 3.0 | fix_normal_random |
|------------------------------|---------------|----------------------------|
|  Total time:                  | 14.133 s   | 14.084 s                   | 
|  Time to initialize:       | 0.002 s     | 0.002 s                     |
|  Time to create:          | 0.129 s     | 0.123 s                     |
|  Time to connect:       | 1.366 s     | 1.382 s                     |
|  Time to presimulate: | 4.028 s     | 3.913 s                     |
|  Time to simulate:      | 7.648 s     | 7.729 s                     |
|  Time to evaluate:      | 0.940 s     | 0.935 s                     |