@jougs 2021-08-20
Ah, OK!

I did not change anything about how errors are caught by the testsuite. IIRC, this was last touched by @heplesser in #1587. However, I am pretty sure we have no mechanism for catching segfaults.