@heplesser 2021-08-21
@niltonlk Thank you for your benchmarks. I cannot reproduce your slow threaded performance. I tested with with current master (062e272) on my MacBook Pro with quad-core i5, compiled with Apple Clang 12.0.5 under OSX 11.5.1. I ran the PD benchmark out of the box. The only change I made was to set `local_num_threads` to 4 for the 1 MPI-4 Thread test.

I obtained the following results from three separate runs (Total time):

- 1 MPI, 1 Thread: 15.074 s, 15.572 s, 15.002 s
- 4 MPI, 1 Thread: 5.936, s 5.722, s 5.718 s
- 1 MPI, 4 Threads: 5.785 s, 5.645 s, 6.225 s

The times for the first two cases are similar to your results, but in my case threaded performance is on par with MPI performance. Could you check your configuration again to see if there may have been something preventing successful threading?